Wrote some tests for the tetraquark contractions, still no bugs found. Moved extraction codes into new extractions folder and moved get_spinmatrix into spinmatrix_ops*

+++ src/HEADERS/polyakov.h
@@ -21,6 +21,6 @@ static_quark( struct spinor *S ,
  */
 double complex
 poly( const struct site *__restrict lat , 
-      int dir ) ;
+      size_t dir ) ;
 
 #endif
